pyproject.toml
[project]
name = "end-to-end-elt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "numpy>=2.3.4",
    "dlt[athena,duckdb,parquet,postgres,workspace]>=1.18.2",
    "faker>=38.0.0",
    "uvicorn>=0.38.0",
    "pydantic[email]>=2.12.4",
    "notebook>=7.4.7",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.44",
    "fastapi[standard]>=0.121.1",
    "bcrypt>=5.0.0",
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "watchdogs>=1.9.9",
    "awswrangler>=3.14.0",
    "boto3>=1.40.70",
]
